# Change latency bucket size for API server metrics