@yue9944882 2018-08-16
@shyamjvs @wojtek-t not sure. but the intention of this pull makes sense to me. whats more it'll no impact on performance.