@shyamjvs 2018-08-16
Changing the buckets may still be fine (I'm not sure), but renaming metrics does sound like a breaking change. Please hold until this is sorted out.

@lavalamp @smarterclayton