@brancz 2018-08-16
We don't have any stability guarantees on metrics and we have broken them in various ways in the past, which doesn't mean we should continue to do so for no reason, but as long as breaking metrics puts them in line with the [official instrumentation guidelines](https://github.com/kubernetes/community/blob/master/contributors/devel/instrumentation.md), I think a change is for the better.

Changing the buckets already breaks the semantics of the metric and as a user of any system I prefer a hard break, rather than a subtle one.