@brancz 2018-08-17
I completely agree that this needs to be widely and clearly communicated.

We have been discussing in sig-instrumentation to do a general overhaul of all the metrics we expose to at least ensure everything is in line with our instrumentation guidelines. What does everyone think of instead of doing one off improvements, instead collecting all the metrics we would like to change and do it in one go? There are numerous violations across the codebase, in terms of metric naming, label naming and non-base units being used, unifying this will give users a better experience working with metrics exposed by Kubernetes.