@smarterclayton 2018-08-17
That seems reasonable if sig driven, and the more changes that happen
together the less likely an individual user is to not notice something
important happened

On Fri, Aug 17, 2018 at 3:45 AM Frederic Branczyk <notifications@github.com>
wrote:

> I completely agree that this needs to be widely and clearly communicated.
>
> We have been discussing in sig-instrumentation to do a general overhaul of
> all the metrics we expose to at least ensure everything is in line with our
> instrumentation guidelines. What does everyone think of instead of doing
> one off improvements, instead collecting all the metrics we would like to
> change and do it in one go? There are numerous violations across the
> codebase, in terms of metric naming, label naming and non-base units being
> used, unifying this will give users a better experience working with
> metrics exposed by Kubernetes.
>
> —
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub
> <https://github.com/kubernetes/kubernetes/pull/67476#issuecomment-413785762>,
> or mute the thread
> <https://github.com/notifications/unsubscribe-auth/ABG_p2zkmV4ucZBc4O_7HardXnvcVbGBks5uRnSGgaJpZM4V_Mwx>
> .
>