@lavalamp 2018-08-17
Making a batch update for the metrics changes sounds good, especially if it comes with an easy-to-understand breakdown of the changes.