@mikkeloscar 2018-08-21
So how do we continue here? Do you want to consider this individual change or do you want to make it part of a batch update?

Regarding the rename and breaking backwards compatibility I totally understand your concerns and as a user of Kubernetes (or any system really) I'm also in favor of stability. However, the fact that my original issue (#63750) didn't get any comments for 3 months and no-one reported a similar issue afaik, suggest to me that not many people are using these metrics at this moment, or if they are, they don't know that they are wrong :)
I'm not suggesting this is a good enough reason to break backwards compatibility, but something to keep in mind.