@matthiasr 2018-09-08
👋 user of these metrics here – in our case most request latencies are longer than 125ms so the existing bucketing is useful, that's why we did not see/comment on #63750.

This is not to say that it could or should not be improved, just verifying that these metrics are being used and communication around updating them is necessary. A major renaming is painful but better than many subtle changes in my opinion.