@wojtek-t 2019-03-11
This has been fixed in the meantime.